+++ assets/index-Bn7SIZ1E.js
@@ -1,1 +1,0 @@
-import{g as r}from"./iframe-YVy6O9s8.js";import{r as o}from"./index-zI4cujTs.js";var t=o();const m=r(t);export{m as H,t as r};