+++ assets/introcard.stories-BoQm0eAi.js
@@ -1,0 +1,5 @@
+import{j as t}from"./iframe-CmpgjNXe.js";import{I as s}from"./index-0o6uPK2A.js";import"./UsersIcon-CPgNwwg_.js";import"./type-Bc9hys3F.js";import"./WebsiteIcon-DC7XU9Lx.js";import"./index-blsVFMDG.js";import"./index-DleQFLpV.js";import"./index-DfQaVWj5.js";import"./OpenIcon-DBsvCFJ4.js";import"./Heading-BwqeFJJP.js";import"./Text-BuvVWvZD.js";const h={title:"PAC Universal Plugin UI/Composite/Intro Card",component:s},r={render:()=>t.jsx("div",{style:{width:"var(--pac-upui-width-sidebar)"},children:t.jsx(s,{})})};var o,i,e;r.parameters={...r.parameters,docs:{...(o=r.parameters)==null?void 0:o.docs,source:{originalSource:`{
+  render: () => <div style={{
+    width: "var(--pac-upui-width-sidebar)"
+  }}><IntroCard /></div>
+}`,...(e=(i=r.parameters)==null?void 0:i.docs)==null?void 0:e.source}}};const C=["Introduction"];export{r as Introduction,C as __namedExportsOrder,h as default};