+++ assets/overview-page.stories-BfQ7bPpO.js
@@ -1,0 +1,34 @@
+import{r as c,j as p}from"./iframe-CmpgjNXe.js";import{O as s}from"./OverviewPage-CHdckNGF.js";import{D as l}from"./DiviCustomFieldHeplerIcon-1hOEmMcI.js";import"./index-DleQFLpV.js";import"./index-Ql0XHf9v.js";import"./Heading-BwqeFJJP.js";import"./Text-BuvVWvZD.js";import"./index-BRZTqdiu.js";import"./index-Dip6cJ9I.js";import"./FilledIcon-DDlOG96s.js";import"./BackIcon-C7xlhmeu.js";import"./type-Bc9hys3F.js";import"./index-H7epRjiP.js";import"./CloseIcon-Dj3JH9Vv.js";import"./PACLogo-DGtyh6ta.js";import"./ChevronRightIcon-DAp352Vg.js";import"./KeyIcon-DGWbnVbZ.js";import"./DocumentationIcon-_4rh2GDI.js";import"./OpenIcon-DBsvCFJ4.js";import"./SupportIcon-DHP7q_ru.js";import"./UsersIcon-CPgNwwg_.js";import"./index-Ck7H4Mnj.js";import"./EyeIcon-CdAFXquu.js";import"./index-BE3P3rPy.js";import"./index-blsVFMDG.js";import"./index-DvaF2tFv.js";import"./index-Dq7h7Pqt.js";import"./index-DcrJ_WHx.js";import"./InfoIcon-D_6c5qcP.js";import"./StopIcon-CfXaxLy2.js";import"./TextLink-Bex_tWyY.js";const W={title:"PAC Universal Plugin UI/Pages/Overview Page",component:s,parameters:{pageLayout:"page"}},e=()=>{const[r,a]=c.useState(!0),n={avatar:"https://lh3.googleusercontent.com/ogw/AF2bZyg0NWhbUjvHEo5VTQIPmq2lgGyO4xDZeaWXPyHq5OknKQ=s32-c-mo",name:"Muhammad Touseef",greeting:"Good afternoon",pluginLogo:l,pluginDescription:"A huge collection of incredibly handy features and tools to make your life easier when working on your Divi website - startup actions, general utilities, interface hacks, custom settings, global styles, learning resources, maintenance tools, code editors, accessibility improvements, alt text fixes, media hacks, local fonts, system report sharing, import/export tools, and so much more!",pluginName:"Divi Assistant",pacPluginNewsApi:"https://dummyjson.com/c/f371-88b4-4d1e-b88e",onPacCardClick:()=>{console.log("PAC Card Clicked")},setCurrentTab:m=>{console.log("Current Tab Set To:",m)},docLink:"https://www.peeayecreative.com/docs/divi-assistant/",supportLink:"https://www.peeayecreative.com/support/",communityLink:"https://www.peeayecreative.com/community/space/divi-assistant/home",showBetaNotice:!0,betaNoticeLink:"https://www.peeayecreative.com/docs/divi-assistant/divi-5/",showReviewCard:r,onReviewCardDismiss:()=>{a(!1)},reviewCardLink:"https://www.elegantthemes.com/marketplace/divi-assistant/reviews/list/1",showLicenseAlert:!0,onLicenseAlertDismiss:()=>{console.log("License Alert Dismissed")}};return p.jsx(s,{...n})};e.__docgenInfo={description:"",methods:[],displayName:"Default"};var t,o,i;e.parameters={...e.parameters,docs:{...(t=e.parameters)==null?void 0:t.docs,source:{originalSource:`() => {
+  const [showReviewCard, setShowReviewCard] = useState(true);
+  const OverviewPageProps = {
+    avatar: "https://lh3.googleusercontent.com/ogw/AF2bZyg0NWhbUjvHEo5VTQIPmq2lgGyO4xDZeaWXPyHq5OknKQ=s32-c-mo",
+    name: "Muhammad Touseef",
+    greeting: "Good afternoon",
+    // pluginLogo: "https://dummyjson.com/image/74",
+    pluginLogo: DiviCustomFieldHeplerIcon,
+    pluginDescription: "A huge collection of incredibly handy features and tools to make your life easier when working on your Divi website - startup actions, general utilities, interface hacks, custom settings, global styles, learning resources, maintenance tools, code editors, accessibility improvements, alt text fixes, media hacks, local fonts, system report sharing, import/export tools, and so much more!",
+    pluginName: "Divi Assistant",
+    pacPluginNewsApi: "https://dummyjson.com/c/f371-88b4-4d1e-b88e",
+    onPacCardClick: () => {
+      console.log("PAC Card Clicked");
+    },
+    setCurrentTab: (tab: string) => {
+      console.log("Current Tab Set To:", tab);
+    },
+    docLink: "https://www.peeayecreative.com/docs/divi-assistant/",
+    supportLink: "https://www.peeayecreative.com/support/",
+    communityLink: "https://www.peeayecreative.com/community/space/divi-assistant/home",
+    showBetaNotice: true,
+    betaNoticeLink: "https://www.peeayecreative.com/docs/divi-assistant/divi-5/",
+    showReviewCard: showReviewCard,
+    onReviewCardDismiss: () => {
+      setShowReviewCard(false);
+    },
+    reviewCardLink: "https://www.elegantthemes.com/marketplace/divi-assistant/reviews/list/1",
+    showLicenseAlert: true,
+    onLicenseAlertDismiss: () => {
+      console.log("License Alert Dismissed");
+    }
+  };
+  return <OverviewPage {...OverviewPageProps} />;
+}`,...(i=(o=e.parameters)==null?void 0:o.docs)==null?void 0:i.source}}};const Z=["Default"];export{e as Default,Z as __namedExportsOrder,W as default};